requirements.txt
appnope==0.1.0
asgiref==3.2.7
backcall==0.1.0
botometer==1.4
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
nltk==3.4.5
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
ParallelDots==3.2.13
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
PySocks==1.7.1
python-dateutil==2.8.1
python-twitter==3.5
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
textblob==0.15.3
traitlets==4.3.3
tweepy==3.8.0
urllib3==1.25.8
wcwidth==0.1.9
whitenoise==5.0.1
